Add a typecast to allow comparing a signed and unsigned int
- mBufferCount is an unsigned int
- Typecast to uint32_t to allow comparing it with index
  which is an int

+++ QCamera2/HAL/QCameraMem.cpp
@@ -998,7 +998,7 @@ camera_memory_t *QCameraStreamMemory::getMemory(int index, bool metadata) const
 *==========================================================================*/
 native_handle_t *qcamera::QCameraVideoMemory::getNativeHandle(uint32_t index, bool metadata)
 {
-    if (index >= mMetaBufCount || (!metadata && index >= mBufferCount)) {
+    if (index >= mMetaBufCount || (!metadata && index >= (uint32_t)mBufferCount)) {
         return NULL;
     }
     return mNativeHandle[index];
